Converting Sessions model into typescript
